actions fixes (#3)
Correct github actions build

Apparently, only setup-uv and uv build are needed in the build stage.  When uv was installing python and syncing the project, it was causing setuptools-scm to increment the version and detect distance from the last tag.

Maybe the .gitignore could have done it too, but I much prefer the simpler github action task.